server/requirements.txt
click==8.0.1
Flask==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.22
protobuf==3.17.3
python-dotenv==0.19.0
six==1.16.0
Werkzeug==2.0.1
